Refactor Index into database
* Pull database, Measurement, Series, ReplicationPolicy, and Shard into database.go to keep things manageable and clean
* Remove the locks from the index method. The server will handle thread safety for database objects
* Move the Index tests into database_test.go and in the influxdb package because we're not exporting database
